Fix issues if the default activtity order settings is used.
In this case, a specific scope filter needs to be added to only include the reshares the user did.
A new function is getting this new needed data.
As the user reshare subnav is used in notifications we also need to edit the query so that we can fetch the the activity of the displayed user others reshared.